Move stdlib_src under test_utils
